Do not jump automatically with :Grep
Since we opened the qf list we can select the result easily anyway
pressing enter, and we avoid opening the buffer if we don't need the
first result.

Additionally, when using the helm quirks autocmd to disable LSP, opening
automatically the file and then the qf list will move the current buffer
from the opened file to the qf list too soon, causing the handler to
attempt to detach the LSP from the qf list buffer instead of the actual
buffer containing the helm template.